requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-chartjs==2.3.0
flower==2.0.1
frozenlist==1.4.0
h11==0.14.0
humanize==4.8.0
idna==3.4
kombu==5.3.2
magic-filter==1.0.12
multidict==6.0.4
outcome==1.2.0
packaging==23.2
Pillow==10.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.3.3
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.2
